feat: extracting logic to custom hooks
